rough sketches for 2 new areas
fence:
player must try to stay on the path; if they fall of they will be transported to the beginning.
possible variation through different paths (e. g. change of level because of size and subsequent enviroment change)

wind:
introduction of wind mechanic (player must block wind or have an adequat size to pass barrier);
at second obstacle player migth be pushed of ledge or lifted to other side (if they are small enough)